requirements.txt
absl-py==2.3.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.61.1
idna==3.11
jax==0.6.2
jaxlib==0.6.2
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.14
ml_dtypes==0.5.4
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
protobuf==4.25.8
pycparser==2.23
pygame==2.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
scipy==1.15.3
six==1.17.0
sounddevice==0.5.3
urllib3==2.6.2
